Move vectorization code to separate file
